Stop-Gap to work around state racing
The alteration of the openssl x509 types has an inconsistent runtime behavior of sometimes executing when we expect it to, and other times completely late to the party after the first set of keys have been generated.

This necessitates the ability to regenerate the keys in the PKI registry, otherwise we might potentially miss out on critcal ssl configuration (which enables apps to work)

This adds a reactor to a transient state 'easyrsa configured' which will forcibly regenerate certificates when this state is reached, and then promptly remove its own state from the runlist.

This causes a chatty re-circulation of the states on the unit, however this is better than the alternative of a racey state.